Update configuration.yaml
Domain level exclusions didn't work because, well, the chatty entities all seem to be sensors, and I can't very well exclude all sensors. So - we need to exclude them specifically, which I now attempt to do.

This will need to be broken-out into a separate file soon. Messy as heck.