Feature health check (#104)
* Initial commit

This is a basic structure, with a very basic hook for pytest, and a callback that seems to have at least invoked today at 6am.....

Aside from that - no idea where the logging is going, etc. So much to do here.

* Changed logging

* Create systemHealthCheck.py

* Update systemHealthCheck.py

Aligned to what I have running on the the Pi at the moment.

* Introduce sonoff platform

off topic, but who cares

* correction

* Reverting sonoff

Have reverted the sonoff definition.

My builds are breaking.

I need to fix the travis build scripts. See:

https://community.home-assistant.io/t/travis-error-with-hacs-ha-community-store/123161

* Fixed indentation issue in appdaemon.yaml

See changes - gave a really confusing error at boot time.

* Fixed DNS test case

cheated, and did it with os.

* Adding simple email notification

This almost certainly won't work on the first try.

* Gets to "Error: unable to send email"

* Starting with jinja2, email packages

The test email is working now.

Required that I use the "less secure" option for the account i am using with gmail. No significant issue with the code on the last commit.

* Results now in message body

with color text

* Added the speed test Test

* Added to-do's

* Params added

Parameterized two of the primitives.

* Added Pi checks

* style correction