Feature presence occupancy (#110)
* initital attempt to change motion sensors

on occupancy

* correction

* Update presenceOccupancy.py

* alarm

* Try just using the GPS

* callback correction

* Change to modelling of Pete's home/not_home

* Update configuration.yaml

* Update presenceOccupancy.py

* Experimenting with owntracks

* removed my half-baked WiFi network detector code

* Initial attempt at state machine

* Corrected syntax error

* Correction to work around param limitation

* Removing the sensor code

* further correction

* Replaced the sensor with an input_select

* updated UI

* Reverting to a sensor

* correction

* Corrected state inconsistency

* minor correction

* correction on the "guest mode" handler

* adding debug

* corrected state string literals

* Corrected the cancel callback call

I used cancel_listen_state. Seems it should be cancel_timer. Should get fewer warnings in the logs now (requires testing).

* defect - alarm callback not being cancelled correctly

I was marked as Just Left, and then went back immediately into Home. And so - the alarm callback cancellation did not trip, because I didn't specify.... I have corrected this bug.

* Fixed bug: stuck in Just Left on boot

On boot with both Persons Away - the state remains in Just Left for > 30 mins. Why? Because I didn't set the timer for the next state transition as part of the initial state setup.

So - i have corrected that.

Let's commit, and see if that fixes things.