Feature delete old backups (#117)
* Code complete

* correction

* most minor of edits

failed to parameter-ize "exec".

Tested on the command line; seems to work.

* Using HACS instead

* Minor cleanup

* Correction to secrets

To align with the production system.

* Attempt to fix the travis build file

Travis build file misaligned now that i am using the custom_component. Attempting to fix it.

* fix attempt - omit pip

* Switch to a later version of python

* Removing deprecated references

* Apparently python 3.8 is the go

resetting this to the value one of the bigger HA deployments is using.

* Speculative commit

trying pip

* Some suggestion it's the base url, that's wrong

https://github.com/tmonck/clean_up_snapshots/issues/1

* Deactivate nest integration

https://github.com/home-assistant/core/issues/53427

There's a problem with one of the libraries associated with the nest integration, which is breaking the build. Bloody thing doesn't work anymore anyway (thanks Google), so let's just suppress it and move on with life.

* adding pip install python-dateutil

It's throwing up on dateutil.

https://github.com/home-assistant/core/issues/28960

Indicates I can clear this by adding a manual installation with pip. I guess we will give that a go.

* Deleting speedtestdotnet

It's now GUI-configured. Killing this to remove the warning message.

* Removed static broadlink config

I have commented out the broadlink config, now that this integration is configured 100% through the GUI.

* Attempt to switch out the old phone

* reverting the group defn

This part is not working. Too ambitious perhaps. Will retry soon.

* attempt to fix